theme2

+++ src/database/migrations/2019_10_17_181523_users_u1.php
@@ -12,28 +12,32 @@ class UsersU1 extends Migration {
 	 */
 	public function up() {
 		Schema::table('users', function (Blueprint $table) {
-			$table->dropColumn('name');
+			if (!Schema::hasColumn('users', 'name') && Schema::hasColumn('users', 'company_id')) {
 
-			$table->unsignedInteger('company_id')->after('id');
-			$table->unsignedInteger('user_type_id')->nullable()->after('company_id');
-			$table->unsignedInteger('entity_id')->nullable()->after('user_type_id');
-			$table->string('first_name', 32)->nullable()->after('entity_id');
-			$table->string('last_name', 32)->nullable()->after('first_name');
-			$table->string('username', 32)->after('last_name');
-			$table->string('mobile_number', 12)->nullable()->after('email');
-			$table->boolean('force_password_reset')->default(0)->after('password');
-			$table->boolean('has_mobile_login')->after('force_password_reset');
-			$table->string('imei', 64)->nullable()->after('has_mobile_login');
-			$table->string('otp', 6)->nullable()->after('imei');
-			$table->string('mpin', 4)->nullable()->after('otp');
-			$table->unsignedInteger('profile_image_id')->nullable()->after('mpin');
-			$table->unsignedInteger('created_by_id')->nullable()->after('remember_token');
-			$table->unsignedInteger('updated_by_id')->nullable()->after('created_by_id');
-			$table->unsignedInteger('deleted_by_id')->nullable()->after('updated_by_id');
-			$table->softdeletes();
+				$table->dropColumn('name');
 
-			$table->foreign('user_type_id')->references('id')->on('configs')->onDelete('CASCADE')->onUpdate('cascade');
-			$table->foreign('profile_image_id')->references('id')->on('attachments')->onDelete('SET NULL')->onUpdate('cascade');
+				$table->unsignedInteger('company_id')->after('id');
+				$table->unsignedInteger('user_type_id')->nullable()->after('company_id');
+				$table->unsignedInteger('entity_id')->nullable()->after('user_type_id');
+				$table->string('first_name', 32)->nullable()->after('entity_id');
+				$table->string('last_name', 32)->nullable()->after('first_name');
+				$table->string('username', 32)->after('last_name');
+				$table->string('mobile_number', 12)->nullable()->after('email');
+				$table->boolean('force_password_reset')->default(0)->after('password');
+				$table->boolean('has_mobile_login')->after('force_password_reset');
+				$table->string('imei', 64)->nullable()->after('has_mobile_login');
+				$table->string('otp', 6)->nullable()->after('imei');
+				$table->string('mpin', 4)->nullable()->after('otp');
+				$table->unsignedInteger('profile_image_id')->nullable()->after('mpin');
+				$table->unsignedInteger('created_by_id')->nullable()->after('remember_token');
+				$table->unsignedInteger('updated_by_id')->nullable()->after('created_by_id');
+				$table->unsignedInteger('deleted_by_id')->nullable()->after('updated_by_id');
+				$table->softdeletes();
+
+				$table->foreign('user_type_id')->references('id')->on('configs')->onDelete('CASCADE')->onUpdate('cascade');
+				$table->foreign('profile_image_id')->references('id')->on('attachments')->onDelete('SET NULL')->onUpdate('cascade');
+			}
+
 		});
 	}
